+++ src/classes_scenes.py
@@ -170,20 +170,19 @@ class GameScene(SceneBase):
         self.list_with_terrain_buttons[0].active = True
         self.current_terrain = terrain_list[0]
 
-
-        # TODO: check and remove
+        # mouse related variables
         self.left_mouse_button_down = False
         self.right_mouse_button_down = False
-        self.left_mouse_button_coord = pygame.mouse.get_pos()
+        self.last_used_tile = 0
 
         # initialize the map
         self.map = Map()
 
         # create initial test trains
         self.dict_with_trains = {
-            1: Train(1, self.map.list_with_tiles[1].coord_world, math.radians(90), RED),
-            2: Train(2, self.map.list_with_tiles[2].coord_world, math.radians(0), ORANGE),
-            3: Train(3, self.map.list_with_tiles[12].coord_world, math.radians(270), BLUE),
+            1: Train(1, self.map.dict_with_tiles[1].coord_world, math.radians(90), RED),
+            2: Train(2, self.map.dict_with_tiles[2].coord_world, math.radians(0), ORANGE),
+            3: Train(3, self.map.dict_with_tiles[12].coord_world, math.radians(270), BLUE),
         }
         self.dict_with_trains[1].movement_path = [8,9,10,11,12]
         self.dict_with_trains[2].movement_path = [3,4,5,16,17,18]
@@ -198,76 +197,60 @@ class GameScene(SceneBase):
         Receive all the events that happened since the last frame.
         Handle all received events.
         """
+        mouse_pos = pygame.mouse.get_pos()
         for event in events:
 
             # mouse button down
             if event.type == pygame.MOUSEBUTTONDOWN:
                 # 1 - left click
                 if event.button == 1:
-                    mouse_coord = pygame.mouse.get_pos()
                     button_was_pressed = False
                     # choose mode
-                    if any([mode_button.is_inside(mouse_coord) for mode_button in self.list_with_mode_buttons]):
+                    if any([mode_button.is_inside(mouse_pos) for mode_button in self.list_with_mode_buttons]):
                         button_was_pressed = True
                         for mode_button in self.list_with_mode_buttons:
-                            if mode_button.check_pressing(mouse_coord):
+                            if mode_button.check_pressing(mouse_pos):
                                 self.current_mode = mode_button.option
                     # choose terrain
                     if self.current_mode == "terrain" and \
-                            any([terrain_button.is_inside(mouse_coord) for terrain_button in self.list_with_terrain_buttons]):
+                            any([terrain_button.is_inside(mouse_pos) for terrain_button in self.list_with_terrain_buttons]):
                         button_was_pressed += True
                         for terrain_button in self.list_with_terrain_buttons:
-                            if terrain_button.check_pressing(mouse_coord):
+                            if terrain_button.check_pressing(mouse_pos):
                                 self.current_terrain = terrain_button.option
 
-                    # add new tile
-                    if not button_was_pressed and self.current_mode == "terrain":
-                        coord_world = screen2world(mouse_coord, self.offset_horizontal, self.offset_vertical, self.scale)
-                        self.map.add_tile(coord_world, self.current_terrain)
+                    if not button_was_pressed:
+                        self.left_mouse_button_down = True
+                        self.right_mouse_button_down = False
 
-                    # add new tile
-                    if not button_was_pressed and self.current_mode == "tracks":
-                        pass
-
-                    # TODO: check and remove
-                    self.left_mouse_button_coord = pygame.mouse.get_pos()
+                # 3 - right click
+                if event.button == 3:
+                    self.right_mouse_button_down = True
                     self.left_mouse_button_down = False
-                    
-                #     # press UI windows (based on notebooks)
-                #     for ui_win in self.list_with_windows:
-                #         self.left_mouse_button_down |= ui_win.press_left(self.dict_with_game_state, self.dict_with_units, self.left_mouse_button_coord)
-
-                #     self.left_mouse_button_down = not self.left_mouse_button_down
-
-                # # 3 - right click
-                # if event.button == 3:
-                #     right_mouse_button_coord = pygame.mouse.get_pos()
-                #     # make_windows_from_right_mouse_button(self.dict_with_units, self.list_with_windows, right_mouse_button_coord, \
-                #     #                             screen2world(right_mouse_button_coord, self.offset_horizontal, self.offset_vertical, self.scale))
 
             # mouse button up
             if event.type == pygame.MOUSEBUTTONUP:
                 # 1 - left click
                 if event.button == 1:
                     self.left_mouse_button_down = False
+                    self.last_used_tile = 0
+                    self.current_used_tile = 0
 
                 # 2 - middle click
                 if event.button == 2:
                     # define new view center
-                    mouse_pos = pygame.mouse.get_pos()
                     self.offset_horizontal -= (mouse_pos[0] - WIN_WIDTH/2) / self.scale
                     self.offset_vertical -= (mouse_pos[1] - WIN_HEIGHT/2) / self.scale
 
-                # # 3 - right click
-                # if event.button == 3:
-                #     # press UI windows (based on slide)
-                #     for ui_win in self.list_with_windows:
-                #         ui_win.press_right(self.map, self.dict_with_units, pygame.mouse.get_pos(), keys_pressed[pygame.K_LCTRL])
+                # 3 - right click
+                if event.button == 3:
+                    self.right_mouse_button_down = False
+                    self.last_used_tile = 0
+                    self.current_used_tile = 0
 
                 # 4 - scroll up
                 if event.button == 4:
                     old_scale = self.scale
-                    # mouse_pos = pygame.mouse.get_pos()
 
                     self.scale *= 2
                     if self.scale >= 4: self.scale = 4
@@ -281,7 +264,6 @@ class GameScene(SceneBase):
                 # 5 - scroll down
                 if event.button == 5:
                     old_scale = self.scale
-                    # mouse_pos = pygame.mouse.get_pos()
 
                     self.scale /= 2
                     # if SCALE <= 0.25: SCALE = 0.25
@@ -320,8 +302,31 @@ class GameScene(SceneBase):
         # move down
         if keys_pressed[pygame.K_DOWN] or keys_pressed[pygame.K_s]:
             self.offset_vertical -= move_speed
-        
 
+
+    # handle the remaining logic related to mouse operation
+        coord_world = screen2world(mouse_pos, self.offset_horizontal, self.offset_vertical, self.scale)
+        coord_id = self.map.world2id(coord_world)
+        current_tile_id = self.map.get_tile_by_coord(coord_id)
+        # adding entities
+        if self.left_mouse_button_down and (not current_tile_id or self.last_used_tile != current_tile_id):
+            # add new tile
+            if self.current_mode == "terrain":
+                self.map.add_tile(coord_id, self.current_terrain)
+            # add new track
+            if self.current_mode == "tracks" and self.last_used_tile and current_tile_id:
+                self.map.add_track(self.last_used_tile, current_tile_id)
+            self.last_used_tile = current_tile_id
+        # removing entities
+        if self.right_mouse_button_down:
+            # remove tile
+            if self.current_mode == "terrain" and current_tile_id:
+                # remove tile from train variables
+                for train_id in self.dict_with_trains:
+                    if current_tile_id in self.dict_with_trains[train_id].movement_path:
+                        self.dict_with_trains[train_id].movement_path = []
+                self.map.remove_tile(current_tile_id)
+        
     def update(self):
         """Game logic for the scene."""