+++ doc/changelog.rst
@@ -1,3 +1,8 @@
+v0.1.2 - 08.07.2024
+------
+* Added mechanism for adding and removing tiles;
+* Added mechanism for adding tracks.
+
 v0.1.1 - 04.07.2024
 ------
 * New system of track definition - not by angles but by tiles' id;